[CRITICAL] [Input] [MacOS] fix timestamp overflow on objC side

+++ native/Avalonia.Native/src/OSX/AvnView.mm
@@ -275,7 +275,7 @@
         delta.Y = [event deltaY];
     }
 
-    uint32 timestamp = static_cast<uint32>([event timestamp] * 1000);
+    uint64_t timestamp = static_cast<uint64_t>([event timestamp] * 1000);
     auto modifiers = [self getModifiers:[event modifierFlags]];
 
     if(type != Move ||
@@ -444,7 +444,7 @@
 
     auto key = s_KeyMap[[event keyCode]];
 
-    uint32_t timestamp = static_cast<uint32_t>([event timestamp] * 1000);
+    uint64_t timestamp = static_cast<uint64_t>([event timestamp] * 1000);
     auto modifiers = [self getModifiers:[event modifierFlags]];
 
     if(_parent != nullptr)
@@ -657,7 +657,7 @@
     
     [self unmarkText];
         
-    uint32_t timestamp = static_cast<uint32_t>([NSDate timeIntervalSinceReferenceDate] * 1000);
+    uint64_t timestamp = static_cast<uint64_t>([NSDate timeIntervalSinceReferenceDate] * 1000);
         
     _lastKeyHandled = _parent->BaseEvents->RawTextInputEvent(timestamp, [text UTF8String]);
